+++ R/tidygraph.R
@@ -17,8 +17,8 @@
 #' library(scgraph)
 #' data("minimal_mesh", package = "silicate")
 #' as_tbl_graph(minimal_mesh)
-#' #library(silicate)  ##not quite
-#' #prim <- PRIMITIVE(as_tbl_graph(minimal_mesh))
+#' library(silicate)  
+#' umodel <- SC(as_tbl_graph(minimal_mesh))
 #' #gibble.PATH <- function(x, ...) {
 #' #inner_join(x[["path"]], x[["path_link_vertex"]] %>% group_by(path) %>% summarize(nrow = n()) ) %>%
 #' #  dplyr::mutate(ncol = 2, type = "MULTILINESTRING")
@@ -28,28 +28,46 @@
 #' #silicate:::build_sf()
 #' ## some kind of round trip
 #' 
-as_tbl_graph.PRIMITIVE <- function(x, ...) {
+as_tbl_graph.SC <- function(x, ...) {
   as_tbl_graph(as.igraph(x), ...)
 }
 
 #' @export 
 #' @name as_tbl_graph
 as_tbl_graph.sf <- function(x, ...) {
-  as_tbl_graph(PRIMITIVE(x), ...)
+  as_tbl_graph(SC(x), ...)
 }
+#' @importFrom silicate sc_object
+sc_object.tbl_graph <- function(x, ...) {
+  tibble::tibble(object_ = silicate::sc_uid(1L))
+}
+#' @importFrom silicate sc_edge
+sc_edge.tbl_graph <- function(x, ...) {
+  out <- igraph::as_edgelist(x)
+  tibble::tibble(.vertex0 = out[,1], .vertex1 = out[, 2], edge_ = sc_uid(nrow(out)))
+}
+#' @importFrom silicate sc_vertex
+sc_vertex.tbl_graph <- function(x, ...) {
+  outm <- igraph::layout.auto(x)
+  edge <- sc_edge(x)
+  out <- tibble::tibble(vertex_ = unique(c(edge$.vertex0, edge$.vertex1)))
+  out$x_ <- outm[,1]
+  out$y_ <- outm[,2]
+  out[c("x_", "y_", "vertex_")]
+}
+#' SC
+#' 
+#' tbl_graph methods for the universal model
+#' @inheritParams silicate::SC
+#' @importFrom silicate SC
+#' @export
+#' @name SC
+SC.tbl_graph <- function(x, ...) {
+    O <- sc_object(x)
+  E <- sc_edge(x)
+  ExO <- tibble::tibble(object_ = O$object_, edge_ = E$edge_)
+  structure(list(object = O, object_link_edge = ExO, edge = E, 
+                 vertex = sc_vertex(x)), class = c("SC", "sc"))
 
-PRIMITIVE.tbl_graph <- function(x, ...) {
-  object = tibble(object = sc_uid(1))
-  vertex = bind_cols(as_tibble(igraph::layout.star(x)) %>% 
-                       rename(x_ = .data$V1, y_ = .data$V2), as_tibble(activate(x, "nodes")))
-  vertex$vertex_ <- sc_uid(nrow(vertex))
-  segment = rename(as_tibble(activate(x, "edges")), .vertex0 = .data$from, .vertex1 = .data$to)
-  segment$.vertex0 <- vertex$vertex_[segment$.vertex0]
-  segment$.vertex1 <- vertex$vertex_[segment$.vertex1]
-  a <- list(object = object, vertex = vertex, segment = segment)
-  attr(a, "join_ramp") <- c("object", "path", "path_link_vertex", "vertex")
-  class(a) <- c("PRIMITIVE", "sc")
-  a
 }
-#prim <- PRIMITIVE(as_tbl_graph(graph))