+++ NEWS.md
@@ -1,5 +1,9 @@
 # scgraph dev
 
+* now using CRAN tidygraph
+
+* updated properly to universal SC model in silicate (0.1.0), PRIMITIVE all removed
+
 * update to silicate, migrated into hypertidy
 
 # scgraph 0.0.1